+ store event data in session storage
Until now, all event data would be lost if the page was reloaded.

Now, all data is stored in the session storage (using 'sessionStorage')
instead of saving in a JS variable.

Every time a new event is created, the following steps are performed
now:

  - Storing new data:
    - Newly-created event data is pushed to a 2D JS array.
    - The 2D array is converted to a single string.
    - The string is stored in the session storage.
  - Showing updated data in the sidebar:
    - The string is retrieved from the session storage.
    - It is converted to a 2D array again and the 2D JS array from
      before is overwritten with this.
    - All previous entries from the sidebar are removed and, using the
      updated data, new entries are added to the sidebar.

Instead of letting 'sessionStorage.setItem()' automatically take care of
converting the 2D array into a string, it is done manually. That is
because the automatic process uses commas as delimiters, which makes it
harder to separate event data later, especially if an event title
contains a comma. Although a hopefully unique-enough sequence of
characters ('#$#') has been used, contingencies should be added in the
future.

The cross button in the sidebar is now active. It can now be used to
remove all event data. A new element has been added to the pop-up: a
'Yes' button, so that the user can be displayed a warning before data
removal. Previous warnings for a 'F5' or a 'CTRL+R' keypress have been
removed, since reloading the page does not remove the data anymore.